feat: added check logic and testing
